--- db/migrate/20180525180219_remove_line_count_from_poems.rb
@@ -1,5 +1,0 @@
-class RemoveLineCountFromPoems < ActiveRecord::Migration[5.1]
-  def change
-    remove_column :poems, :linecount, :integer
-  end
-end
