+++ db/schema.rb
@@ -10,7 +10,7 @@
 #
 # It's strongly recommended that you check this file into your version control system.
 
-ActiveRecord::Schema.define(version: 20180525180219) do
+ActiveRecord::Schema.define(version: 20180525140226) do
 
   # These are extensions that must be enabled in order to support this database
   enable_extension "plpgsql"
@@ -24,6 +24,7 @@ ActiveRecord::Schema.define(version: 20180525180219) do
   create_table "poems", force: :cascade do |t|
     t.string "title"
     t.string "lines"
+    t.integer "linecount"
     t.datetime "created_at", null: false
     t.datetime "updated_at", null: false
   end
