+++ db/seeds.rb
@@ -5,6 +5,3 @@
 #
 #   movies = Movie.create([{ name: 'Star Wars' }, { name: 'Lord of the Rings' }])
 #   Character.create(name: 'Luke', movie: movies.first)
-
-rani = Author.create(name:'Rani')
-henry = Author.create(name:'Henry')
